+++ examples/next-14x/app/@tabs/details/page.tsx
@@ -1,9 +1,0 @@
-import { Text } from '@vtex/shoreline'
-
-export default function Page() {
-  return (
-    <main>
-      <Text>Details tab</Text>
-    </main>
-  )
-}