+++ packages/docs/pages/components/collection/_meta.ts
@@ -1,0 +1,5 @@
+export default {
+  'best-practices': 'Best Practices',
+  'collection-view': 'CollectionView',
+  'collection-row': 'CollectionRow',
+}