+++ packages/docs/pages/components/empty-state/_meta.ts
@@ -1,0 +1,4 @@
+export default {
+  'empty-state-actions': 'EmptyStateActions',
+  'empty-state-illustration': 'EmptyStateIllustration',
+}