+++ packages/docs/pages/components/field/_meta.ts
@@ -1,0 +1,6 @@
+export default {
+  'field-char-counter': 'FieldCharCounter',
+  'field-description': 'FieldDescription',
+  'field-error': 'FieldError',
+  'field-provider': 'FieldProvider',
+}