+++ packages/docs/pages/components/filter/_meta.ts
@@ -1,0 +1,13 @@
+export default {
+  'best-practices': 'Best Practices',
+  'filter-apply': 'FilterApply',
+  'filter-clear': 'FilterClear',
+  'filter-item': 'FilterItem',
+  'filter-item-check': 'FilterItemCheck',
+  'filter-list': 'FilterList',
+  'filter-list-skeleton': 'FilterListSkeleton',
+  'filter-popover': 'FilterPopover',
+  'filter-provider': 'FilterProvider',
+  'filter-trigger': 'FilterTrigger',
+  'filter-value': 'FilterValue',
+}