+++ packages/docs/pages/components/menu/_meta.ts
@@ -1,0 +1,8 @@
+export default {
+  'best-practices': 'Best Practices',
+  'menu-item': 'MenuItem',
+  'menu-popover': 'MenuPopover',
+  'menu-provider': 'MenuProvider',
+  'menu-separator': 'MenuSeparator',
+  'menu-trigger': 'MenuTrigger',
+}