+++ packages/docs/pages/components/page/_meta.ts
@@ -1,0 +1,6 @@
+export default {
+  'best-practices': 'Best Practices',
+  'page-content': 'PageContent',
+  'page-header': 'PageHeader',
+  'page-heading': 'PageHeading',
+}