+++ packages/docs/pages/components/table/_meta.ts
@@ -1,0 +1,11 @@
+export default {
+  'tanstack-table': '@tanstack/react-table',
+  'best-practices': 'Best Practices',
+  'figma-usage': 'Figma Usage',
+  'table-body': 'TableBody',
+  'table-cell': 'TableCell',
+  'table-header': 'TableHeader',
+  'table-header-cell': 'TableHeaderCell',
+  'table-row': 'TableRow',
+  'table-sort-indicator': 'TableSortIndicator',
+}